### 2025-08-30 634701f
add arXiv DOIs
