+++ paper/paper.bib
@@ -42,7 +42,8 @@
   month={02},
   eprint={2402.16669},
   eprinttype={arxiv},
-  eprintclass={math.NA}
+  eprintclass={math.NA},
+  doi={10.48550/arXiv.2402.16669}
 }
 
 @article{ranocha2025structure,
@@ -81,7 +82,8 @@
   month={08},
   eprint={2508.04112},
   eprinttype={arxiv},
-  eprintclass={math.NA}
+  eprintclass={math.NA},
+  doi={10.48550/arXiv.2508.04112}
 }
 
 @article{mattsson2014diagonal,
