### 2025-09-03 c7a450d
added readme to data folder
