### 2025-09-03 5fe37f4
Apply suggestions from code review
Co-authored-by: Joshua Lampert <51029046+JoshuaLampert@users.noreply.github.com>
