### 2025-09-03 807f0d8
made reame more verbose and added info to docstring
