### 2025-09-03 4ceaff1
Update src/experimental_data.jl
Co-authored-by: Joshua Lampert <51029046+JoshuaLampert@users.noreply.github.com>
