### 2025-09-18 b45d1e0
WIP: energy-conserving discretization for HyperbolicSainteMarieEquations1D with periodic BCs and flat bathymetry
