+++ src/equations/hyperbolic_sainte_marie_1d.jl
@@ -1,0 +1,290 @@
+struct HyperbolicSainteMarieEquations1D{Bathymetry <: BathymetryFlat,
+                                        RealT <: Real} <:
+       AbstractSainteMarieEquations{1, 5}
+    bathymetry_type::Bathymetry # type of bathymetry
+    gravity::RealT # gravitational acceleration
+    eta0::RealT # constant still-water surface
+    alpha::RealT # hyperbolic relaxation parameter (→ ∞ for Sainte-Marie)
+    c_squared::RealT # c^2 = alpha^2 * g * eta0
+end
+
+function HyperbolicSainteMarieEquations1D(; bathymetry_type = bathymetry_flat,
+                                           gravity,
+                                           eta0,
+                                           alpha)
+    c_squared = alpha^2 * gravity * eta0
+    HyperbolicSainteMarieEquations1D(bathymetry_type, gravity, eta0, alpha, c_squared)
+end
+
+# TODO: What is easier - hv or v as main computational variable?
+function varnames(::typeof(prim2prim), ::HyperbolicSainteMarieEquations1D)
+    return ("η", "v", "D", "w", "p")
+end
+function varnames(::typeof(prim2cons), ::HyperbolicSainteMarieEquations1D)
+    return ("h", "hv", "b", "hw", "hp")
+end
+
+"""
+    prim2phys(q, equations::HyperbolicSainteMarieEquations1D)
+
+Return the physical variables ``\\eta, v, D`` used also by the
+[`SainteMarieEquations1D`](@ref) from the main variables
+`q` for the [`HyperbolicSainteMarieEquations1D`](@ref).
+"""
+function prim2phys(q, ::HyperbolicSainteMarieEquations1D)
+    eta, v, D = q
+    return SVector(eta, v, D)
+end
+
+varnames(::typeof(prim2phys), ::HyperbolicSainteMarieEquations1D) = ("η", "v", "D")
+
+is_hyperbolic_appproximation(::HyperbolicSainteMarieEquations1D) = Val{true}()
+
+function hyperbolic_approximation_limit(equations::HyperbolicSainteMarieEquations1D)
+    (; bathymetry_type, gravity, eta0) = equations
+    return SainteMarieEquations1D(bathymetry_type, gravity, eta0)
+end
+
+# compute the auxiliary variables from the main ones
+# using the default initialization
+function set_approximation_variables!(q, mesh,
+                                      equations::HyperbolicSainteMarieEquations1D,
+                                      solver)
+    (; D1) = solver
+    eta, v, D, w, p = q.x
+
+    # This is not true but just a simple initialization used also in
+    # at least one experiment of Escalante, Dumbser, Castro (2019).
+    @.. p = 0
+
+    # w ≈ -(1/2) h v_x
+    # with h = (h + b) + (eta0 - b) - eta0
+    mul!(w, D1, v)
+    @.. w = -0.5 * (eta + D - equations.eta0) * w
+    if !(equations.bathymetry_type isa BathymetryFlat)
+        error("Non-flat bathymetry is not supported")
+    end
+
+    return nothing
+end
+
+# TODO: other methods
+
+function create_cache(mesh, equations::HyperbolicSainteMarieEquations1D,
+                      solver, initial_condition,
+                      boundary_conditions::BoundaryConditionPeriodic,
+                      RealT, uEltype)
+    # We use `DiffCache` from PreallocationTools.jl to enable automatic/algorithmic differentiation
[…]
