### 2025-09-18 526b575
TODO notes for development
