+++ examples/hyperbolic_sainte_marie_1d/hyperbolic_sainte_marie_conservation.jl
@@ -19,7 +19,7 @@ julia> integrals(analysis_callback).waterheight_total |> x -> (x[end] - x[1]) / 
 0.0
 
 julia> integrals(analysis_callback).momentum |> x -> (x[end] - x[1]) / x[1]
-2.9432057444810533e-15
+2.9432057444810533e-15 # for flat bathymetry
 
 julia> integrals(analysis_callback).entropy_modified |> x -> (x[end] - x[1]) / x[1]
 -1.064590038909088e-15
@@ -28,7 +28,7 @@ julia> integrals(analysis_callback).entropy_modified |> x -> (x[end] - x[1]) / x
 ###############################################################################
 # Semidiscretization of the hyperbolic Sainte-Marie equations
 
-equations = HyperbolicSainteMarieEquations1D(bathymetry_type = bathymetry_flat,
+equations = HyperbolicSainteMarieEquations1D(bathymetry_type = bathymetry_mild_slope,
                                              gravity = 9.81,
                                              eta0 = 1.0,
                                              alpha = 3.0)
@@ -38,7 +38,7 @@ function initial_condition_conservation_test(x, t,
                                              mesh)
     eta = 1 + exp(-x^2)
     v = 1.0e-2 # set this to zero to test a directional bias
-    b = 0.0 # FIXME: 0.25 * cospi(x / 75)
+    b = 0.25 * cospi(x / 75)
 
     # We use the feature that we can only return the physical variables
     # used by the `hyperbolic_approximation_limit`, i.e., the
