### 2025-09-29 f16c15f
define fallback set_approximation_variables! for JET.jl
