+++ test/Project.toml
@@ -2,7 +2,6 @@
 Aqua = "4c88cf16-eb10-579e-8560-4a9242c79595"
 ExplicitImports = "7d51a73a-1435-4ff3-83d9-f097790105c7"
 ForwardDiff = "f6369f11-7733-5829-9624-2563aa707210"
-JET = "c3a54625-cd67-489e-a8e7-0a5a0ff4e31b"
 OrdinaryDiffEqLowStorageRK = "b0944070-b475-4768-8dec-fb6eb410534d"
 OrdinaryDiffEqRosenbrock = "43230ef6-c299-4910-a778-202eb28ce4ce"
 OrdinaryDiffEqTsit5 = "b1df2697-797e-41e3-8120-5422d3b24e4a"
@@ -18,7 +17,6 @@ TrixiTest = "0a316866-cbd0-4425-8bcb-08103b2c1f26"
 Aqua = "0.7, 0.8"
 ExplicitImports = "1.0.1"
 ForwardDiff = "0.10.36, 1"
-JET = "0.10"
 OrdinaryDiffEqLowStorageRK = "1.1"
 OrdinaryDiffEqRosenbrock = "1.9"
 OrdinaryDiffEqTsit5 = "1.1"
