+++ .github/workflows/JET.yml
@@ -1,4 +1,4 @@
-name: CI
+name: JET
 on:
   push:
     branches:
