### 2025-10-08 3b237f2
Apply suggestions from code review
