+++ .github/workflows/JET.yml
@@ -48,7 +48,7 @@ jobs:
       fail-fast: false
       matrix:
         version:
-          - '~1.12.0-0'
+          - '1.12'
         os:
           - ubuntu-latest
     steps:
