### 2025-10-23 038fdef
Do not run benchmarks on non-code changes
Added paths to ignore for benchmark workflow.
