+++ .github/workflows/benchmark.yml
@@ -2,6 +2,19 @@ name: Benchmark this PR
 on:
   pull_request_target:
     branches: [ main ]  # change to your default branch
+    paths-ignore:
+      - 'CITATION.bib'
+      - 'LICENSE.md'
+      - 'README.md'
+      - '.zenodo.json'
+      - '.github/workflows/benchmark.yml'
+      - '.github/workflows/CompatHelper.yml'
+      - '.github/workflows/Documenter.yml'
+      - '.github/workflows/Format-check.yml'
+      - '.github/workflows/TagBot.yml'
+      - '.github/workflows/SpellCheck.yml'
+      - 'benchmark/**'
+      - 'docs/**'
 permissions:
   pull-requests: write    # action needs to post a comment
 
