### 2025-10-23 c87f789
Update .github/workflows/benchmark.yml
Co-authored-by: Joshua Lampert <51029046+JoshuaLampert@users.noreply.github.com>
