+++ .github/workflows/benchmark.yml
@@ -7,7 +7,6 @@ on:
       - 'LICENSE.md'
       - 'README.md'
       - '.zenodo.json'
-      - '.github/workflows/benchmark.yml'
       - '.github/workflows/CompatHelper.yml'
       - '.github/workflows/Documenter.yml'
       - '.github/workflows/Format-check.yml'
