### 2025-10-23 bb7fd86
Update .github/workflows/benchmark.yml
Co-authored-by: Joshua Lampert <51029046+JoshuaLampert@users.noreply.github.com>
