+++ .github/workflows/benchmark.yml
@@ -12,7 +12,6 @@ on:
       - '.github/workflows/Format-check.yml'
       - '.github/workflows/TagBot.yml'
       - '.github/workflows/SpellCheck.yml'
-      - 'benchmark/**'
       - 'docs/**'
 permissions:
   pull-requests: write    # action needs to post a comment
