### 2025-11-03 8d53fd1
fixed some issues found by JET.jl
