### 2025-11-03 9168ccf
always unpack psi in if conditions
