### 2025-11-06 #8894 @jatkinson1000
@navidcy JOSS policy is for a minimum of two reviewers per paper, but a preference for three. It is not unheard of for reviewers to vanish after agreeing to review, and this can cause delays with finding reviewers at a later stage once review is underway. If the author wishes to start now with two then we can do so - @JoshuaLampert

I will do a last request for reviewers and if no success we can make a start.
