### 2025-11-06 #8894 @jatkinson1000
Hi there @Sush1090, @alexfikl, @rahulkhorana

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a Julia package that implements a number of numerical methods for shallow-water waves.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
