### 2025-11-06 #8894 @pnavaro
Hi @jatkinson1000, I discussed with my colleague @vybduchene. Vincent is a mathematician, he is familiar with these numerical methods and he is the main developer of [WaterWaves1D.jl](https://github.com/WaterWavesModels/WaterWaves1D.jl), but he does not feel qualified to judge the software quality of the project. However, I agreed to help him. We are from the same lab. Do we only need one name, or can we both be involved?
