### 2025-11-06 #9361 @editorialbot
[REVIEW]: DispersiveShallowWater.jl: A Julia library of structure-preserving numerical methods for dispersive wave equations
**Submitting author:** @JoshuaLampert (<a href="http://orcid.org/0009-0007-0971-6709">Joshua Lampert</a>)
**Repository:** https://github.com/NumericalMathematics/DispersiveShallowWater.jl
**Branch with paper.md** (empty if default branch): paper-2025-joss
**Version:** v0.8.7
