### 2025-11-07 JoshuaLampert/DispersiveShallowWater.jl#249 @Alexander-Barth
link in README.md
As you probably know, I am reviewing your JOSS submission :-)
I am wondering if in README.md, the link (in the Usage section)

https://numericalmathematics.github.io/DispersiveShallowWater.jl/stable/overview/#visualize_results

should rather be this one?

https://numericalmathematics.github.io/DispersiveShallowWater.jl/stable/basic_example/#visualize_results

ref: https://github.com/openjournals/joss-reviews/issues/9361
