### 2025-11-08 a087a65
add references to additional Julia packages
