+++ paper/paper.bib
@@ -332,6 +332,37 @@
   doi = {10.21105/joss.02018}
 }
 
+@article{GeophysicalFlowsJOSS,
+  doi = {10.21105/joss.03053},
+  url = {https://doi.org/10.21105/joss.03053},
+  year = {2021},
+  publisher = {The Open Journal},
+  volume = {6},
+  number = {60},
+  pages = {3053},
+  author = {Navid C. Constantinou and Gregory LeClaire Wagner and Lia Siegelman and Brodie C. Pearson and André Palóczy},
+  title = {{GeophysicalFlows.jl: Solvers for geophysical fluid dynamics problems in periodic domains on CPUs \& GPUs}},
+  journal = {Journal of Open Source Software}
+}
+
+@article{paprota2025spectralwaves,
+  title={A fully spectral framework for nonlinear water waves propagating over topography},
+  author={Paprota, Maciej},
+  year={2025},
+  journal={Coastal Engineering},
+  volume={200},
+  issn={0378-3839},
+  howpublished={\url{https://github.com/mcpaprota/SpectralWaves.jl}},
+  doi={10.1016/j.coastaleng.2025.104759}
+}
+
+@misc{duchenewaterwaves1d,
+  title={WaterWaves.jl},
+  author={Duch{\^e}ne, Vincent and Navaro, Pierre},
+  howpublished={\url{https://github.com/WaterWavesModels/WaterWaves1D.jl}},
+  doi={10.5281/zenodo.7142921}
+}
+
 
 @article{ketcheson2019relaxation,
   title = {
