+++ paper/paper.bib
@@ -357,7 +357,7 @@
 }
 
 @misc{duchenewaterwaves1d,
-  title={WaterWaves.jl},
+  title={WaterWaves1D.jl},
   author={Duch{\^e}ne, Vincent and Navaro, Pierre},
   howpublished={\url{https://github.com/WaterWavesModels/WaterWaves1D.jl}},
   doi={10.5281/zenodo.7142921}
