### 2025-11-08 #9361 @giordano
> We have added a few references of similar Julia packages that we learned about in the meantime.

I think also https://github.com/milankl/ShallowWaters.jl would be in scope?
