### 2025-11-10 f5c9954
add reference to ShallowWaters.jl
