+++ paper/paper.bib
@@ -363,6 +363,13 @@
   doi={10.5281/zenodo.7142921}
 }
 
+@misc{klowershallowwaters,
+  title={ShallowWaters.jl - A type-flexible 16-bit shallow water model},
+  author={Klöwer, Milan},
+  howpublished={\url{https://github.com/milankl/ShallowWaters.jl}},
+  doi={10.5281/zenodo.3626102}
+}
+
 
 @article{ketcheson2019relaxation,
   title = {
