### 2025-11-10 5c1cfad
update paper references
