### 2025-11-10 dda723f
use doi.org links
