### 2025-11-10 441b6a3
update paper reference
