+++ paper/paper.bib
@@ -34,16 +34,19 @@
   doi = {10.1007/s10543-023-00992-w}
 }
 
-@online{lampert2024structure,
-  title={Structure-Preserving Numerical Methods for Two Nonlinear Systems of
-         Dispersive Wave Equations},
-  author={Lampert, Joshua and Ranocha, Hendrik},
-  year={2024},
-  month={02},
-  eprint={2402.16669},
-  eprinttype={arxiv},
-  eprintclass={math.NA},
-  doi={10.48550/arXiv.2402.16669}
+@article{lampert2025structure,
+  title = {Structure-preserving numerical methods for two nonlinear systems of
+         dispersive wave equations},
+  author = {Lampert, Joshua and Ranocha, Hendrik},
+  journal = {Computational Science and Engineering},
+  volume = {2},
+  number = {2},
+  year = {2025},
+  month = {10},
+  doi = {10.1007/s44207-025-00006-3},
+  eprint = {2402.16669},
+  eprinttype = {arxiv},
+  eprintclass = {math.NA}
 }
 
 @article{ranocha2025structure,
