### 2025-11-18 JoshuaLampert/DispersiveShallowWater.jl#256 @pnavaro
the default example is physically unrealistic
Regarding the default example `BBM_BBM_1D_BASIC`. It seems unfortunate that this example produces negative water heights, which is physically unrealistic.

```julia
using DispersiveShallowWater
include(default_example())
using Plots
plot(semi => sol)
```

[JOSS review #9361](https://github.com/openjournals/joss-reviews/issues/9361)
