### 2025-11-18 NumericalMathematics/DispersiveShallowWater.jl#257 @pnavaro
Add more explanations for all callback analysis quantities
The `callback_analysis` feature is very helpful for analyzing model precision, but we could not find explanations for some quantities such as ∫U for instance. These quantities could be clarified in the documentation of the equations.

[JOSS review #9361](https://github.com/openjournals/joss-reviews/issues/9361)
