### 2025-11-18 JoshuaLampert/DispersiveShallowWater.jl#258 @pnavaro
Svaerd-Kalisch model
The implemented Svaerd-Kalisch model differs from the model in the published paper (JCP - equations 11-12). This model has 3 parameters, and when α or γ are non-zero, it is not suitable for modeling water wave propagation. This can be observed by noting that the phase velocity is not an even function of the wave number. As a side note, the convention for computing this quantity for negative wave numbers is not consistent across all models (Euler and Svaerd-Kalisch models are opposite to all other models). We suggest adding a note in the documentation to clearly indicate the parameter restrictions and warn users about the configurations that are not appropriate.

```julia
using DispersiveShallowWater
using Plots

eta0 = 0.0
reference_height() = 0.8
h0 = reference_height()
g = 9.81
disp_rel = LinearDispersionRelation(h0)
k = -5.0:0.01:5.0

bbm = BBMEquation1D(; gravity = g, eta0 = eta0, D = h0)
c_bbm = wave_speed.(disp_rel, bbm, k; normalize = true)
plot(k, c_bbm, label = "BBM")

sk = SvaerdKalischEquations1D(; gravity = g, eta0 = eta0,
                              alpha = 0.0, beta = 0.2308939393939394, gamma = 0.04034343434343434)
c_sk = wave_speed.(disp_rel, sk, k; normalize = true)
plot!(k, abs.(c_sk), label = "Svärd-Kalisch")
```

[JOSS review #9361](https://github.com/openjournals/joss-reviews/issues/9361)
