### 2025-11-18 NumericalMathematics/DispersiveShallowWater.jl#254 @JoshuaLampert
Thanks for raising the issue. Yes, this is a bit of a naming issue. In DispersiveShallowWater.jl the initial condition can be exploited as analytical solution by using the `t` argument if an analytical solution is known. This is explained a bit in https://numericalmathematics.github.io/DispersiveShallowWater.jl/stable/basic_example/#visualize_results. However, I agree it should at least be also made clear in https://numericalmathematics.github.io/DispersiveShallowWater.jl/stable/plotting/. Maybe it's even worth to add another plot argument `plot_analytical`, which does what the current `plot_initial` does and let `plot_initial` plot the initial condition (which would be breaking). However, this can also be confusing for the case where no analytical solution is given as one might expect the new argument `plot_analytical` still plots an analytical solution. What do you think, @ranocha?
