### 2025-11-18 NumericalMathematics/DispersiveShallowWater.jl#255 @JoshuaLampert
I am not sure I understand the issue. Executing the script above, I see a plot, where the y-axis label is `η`. In our notation explained [here](https://numericalmathematics.github.io/DispersiveShallowWater.jl/stable/overview/#Variable-Descriptions), η is denoted "Total water height", which is also what the function `waterheight_total` refers to. In my understanding this quantity describes the same as the "surface elevation". Do I understand correctly that you would expect the surface elevation and total water height describe two different quantities? I yes, what would you call "total water height"?
