### 2025-11-18 NumericalMathematics/DispersiveShallowWater.jl#255 @pnavaro
For me, surface elevation measures the deviation from a reference and total water height measures the absolute water column depth. The first can be negative and the second is always positive.
