### 2025-11-18 NumericalMathematics/DispersiveShallowWater.jl#255 @JoshuaLampert
I see. Maybe our nomenclature is not the best, but we use it all over the place (I hope consistently) including published articles. So I would be hesitant to change it. I hope the figure in https://numericalmathematics.github.io/DispersiveShallowWater.jl/stable/overview/#Variable-Descriptions at least explains our notation and naming.
