### 2025-11-18 722d4c0
ckarify variable names in some docstrings
