+++ src/equations/equations.jl
@@ -142,7 +142,7 @@ end
 """
     waterheight_total(q, equations)
 
-Return the total waterheight of the primitive variables `q` for a given set of
+Return the total waterheight ``\\eta`` of the primitive variables `q` for a given set of
 `equations`, i.e., the [`waterheight`](@ref) ``h`` plus the
 [`bathymetry`](@ref) ``b``.
 
@@ -175,7 +175,7 @@ varnames(::typeof(waterheight), equations) = ("h",)
 """
     velocity(q, equations)
 
-Return the velocity of the primitive variables `q` for a given set of
+Return the velocity ``v`` of the primitive variables `q` for a given set of
 `equations`.
 
 `q` is a vector of the primitive variables at a single node, i.e., a vector
@@ -190,7 +190,7 @@ varnames(::typeof(velocity), equations) = ("v",)
 """
     momentum(q, equations)
 
-Return the momentum/discharge of the primitive variables `q` for a given set of
+Return the momentum/discharge ``hv`` of the primitive variables `q` for a given set of
 `equations`, i.e., the [`waterheight`](@ref) times the [`velocity`](@ref).
 
 `q` is a vector of the primitive variables at a single node, i.e., a vector
@@ -230,7 +230,7 @@ still_waterdepth(q, equations::AbstractEquations{1, 1}) = equations.D
 """
     bathymetry(q, equations)
 
-Return the bathymetry of the primitive variables `q` for a given set of
+Return the bathymetry ``b`` of the primitive variables `q` for a given set of
 `equations`.
 
 `q` is a vector of the primitive variables at a single node, i.e., a vector
@@ -268,12 +268,12 @@ end
 """
     energy_total(q, equations)
 
-Return the total energy of the primitive variables `q` for a given set of
+Return the total energy ``e`` of the primitive variables `q` for a given set of
 `equations`. For all [`AbstractShallowWaterEquations`](@ref), the total
 energy is given by the sum of the kinetic and potential energy of the
 shallow water subsystem, i.e.,
 ```math
-\\frac{1}{2} h v^2 + \\frac{1}{2} g \\eta^2
+e(\\eta, v) = \\frac{1}{2} h v^2 + \\frac{1}{2} g \\eta^2
 ```
 in 1D, where ``h`` is the [`waterheight`](@ref),
 ``\\eta = h + b`` the [`waterheight_total`](@ref),
@@ -294,7 +294,7 @@ varnames(::typeof(energy_total), equations) = ("e_total",)
 """
     entropy(q, equations)
 
-Return the entropy of the primitive variables `q` for a given set of
+Return the entropy ``U`` of the primitive variables `q` for a given set of
 `equations`. For all [`AbstractShallowWaterEquations`](@ref), the `entropy`
 is just the [`energy_total`](@ref).
 
@@ -311,7 +311,7 @@ varnames(::typeof(entropy), equations) = ("U",)
 """
     energy_total_modified(q_global, equations, cache)
 
-Return the modified total energy of the primitive variables `q_global` for the
+Return the modified total energy ``\\hat e`` of the primitive variables `q_global` for the
 `equations`. This modified total energy is a conserved quantity and can
 contain additional terms compared to the usual [`energy_total`](@ref).
 For example, for the [`SvaerdKalischEquations1D`](@ref) and the
@@ -331,25 +331,25 @@ Internally, this function allocates a vector for the output and
 calls [`DispersiveShallowWater.energy_total_modified!`](@ref).
 """
 function energy_total_modified(q_global, equations::AbstractEquations, cache)
-    e = similar(q_global.x[begin])
-    return energy_total_modified!(e, q_global, equations, cache)
[…]
