+++ src/equations/serre_green_naghdi_1d.jl
@@ -59,7 +59,8 @@ References for the Serre-Green-Naghdi system can be found in
 The semidiscretization implemented here conserves
 - the total water mass (integral of ``h``) as a linear invariant
 - the total momentum (integral of ``h v``) as a nonlinear invariant if the bathymetry is constant
-- the total modified energy
+- the total modified entropy/energy (integral of ``\hat U``/``\hat e``), which is called here
+  [`entropy_modified`](@ref) and [`energy_total_modified`](@ref)
 
 for periodic boundary conditions (see Ranocha and Ricchiuto (2025)).
 Additionally, it is well-balanced for the lake-at-rest stationary solution, see
@@ -784,7 +785,7 @@ end
 # Discretization that conserves
 # - the total water mass (integral of ``h``) as a linear invariant
 # - the total momentum (integral of ``h v``) as a nonlinear invariant for flat bathymetry
-# - the total modified energy
+# - the total modified entropy/modified energy (integral of ``\hat U``/``\hat e``) as a nonlinear invariant
 # for periodic boundary conditions, see
 # - Hendrik Ranocha and Mario Ricchiuto (2025)
 #   Structure-Preserving Approximations of the Serre-Green-Naghdi
@@ -1352,22 +1353,22 @@ end
 
 # The modified entropy/energy takes the whole `q` for every point in space
 """
-    DispersiveShallowWater.energy_total_modified!(e, q_global, equations::SerreGreenNaghdiEquations1D, cache)
+    DispersiveShallowWater.energy_total_modified!(e_mod, q_global, equations::SerreGreenNaghdiEquations1D, cache)
 
-Return the modified total energy `e` of the primitive variables `q_global` for the
+Return the modified total energy ``\\hat e`` of the primitive variables `q_global` for the
 [`SerreGreenNaghdiEquations1D`](@ref).
 It contains an additional term containing a
 derivative compared to the usual [`energy_total`](@ref) modeling
-non-hydrostatic contributions. The `energy_total_modified`
+non-hydrostatic contributions. The [`energy_total_modified`](@ref)
 is a conserved quantity (for periodic boundary conditions).
 
 For a [`bathymetry_flat`](@ref) the total modified energy is given by
 ```math
-\\frac{1}{2} g \\eta^2 + \\frac{1}{2} h v^2 + \\frac{1}{6} h^3 v_x^2.
+\\hat e(\\eta, v) = \\frac{1}{2} g \\eta^2 + \\frac{1}{2} h v^2 + \\frac{1}{6} h^3 v_x^2.
 ```
 For a [`bathymetry_mild_slope`](@ref) the total modified energy is given by
 ```math
-\\frac{1}{2} g \\eta^2 + \\frac{1}{2} h v^2 + \\frac{1}{6} h (-h v_x + 1.5 v b_x)^2.
+\\hat e(\\eta, v) = \\frac{1}{2} g \\eta^2 + \\frac{1}{2} h v^2 + \\frac{1}{6} h (-h v_x + 1.5 v b_x)^2.
 ```
 For a [`bathymetry_variable`](@ref) the total modified energy has the additional term
 ```math
@@ -1379,7 +1380,7 @@ For a [`bathymetry_variable`](@ref) the total modified energy has the additional
 
 See also [`energy_total_modified`](@ref).
 """
-function energy_total_modified!(e, q_global,
+function energy_total_modified!(e_mod, q_global,
                                 equations::SerreGreenNaghdiEquations1D,
                                 cache)
     # unpack physical parameters and SBP operator `D1`
[…]
