### 2025-11-18 49ec5b5
properly distinguish water height and total water height
