### 2025-11-18 6570b32
math rendering
