+++ dev/search_index.js
@@ -1,0 +1,3 @@
+var documenterSearchIndex = {"docs":
+[{"location":"dingemans/#dingemans","page":"Experimental Data","title":"Dingemans Experiment","text":"The Dingemans experiment provides a classic benchmark for validating dispersive shallow water models against experimental data. This experiment, conducted by Dingemans in 1994 , simulates waves generated by a wave maker that propagate over a varying underwater topography (bathymetry). See more at [Dingemans1994] and [Dingemans1997]","category":"section"},{"location":"dingemans/#Experimental-Setup","page":"Experimental Data","title":"Experimental Setup","text":"The experiment features waves with a small amplitude (A = 0.02) that encounter a trapezoidal bathymetry profile. The bottom topography starts flat, then rises to form an underwater \"hill\" before descending back to the original depth. This setup tests how well different dispersive wave models can capture the complex wave transformations that occur when waves interact with varying bottom topography.\n\nThe bathymetry profile consists of:\n\nA flat section from the wave maker to x = 11.01\nA linearly increasing slope from x = 11.01 to x = 23.04 (maximum height of 0.6)\nA flat plateau from x = 23.04 to x = 27.04\nA linearly decreasing slope from x = 27.04 to x = 33.07\nA flat section beyond x = 33.07\n\nThis configuration allows researchers to study wave shoaling (changes in wave characteristics due to depth variations), which is crucial for understanding coastal wave dynamics.","category":"section"},{"location":"dingemans/#Numerical-Simulation","page":"Experimental Data","title":"Numerical Simulation","text":"Let's implement the Dingemans experiment and compare the performance of different dispersive shallow water models available in DispersiveShallowWater.jl.\n\nFirst, we load the necessary packages:\n\nusing DispersiveShallowWater, SummationByPartsOperators, SparseArrays, OrdinaryDiffEqTsit5, Plots\n\nNext, we set up the different equation systems we want to compare:\n\n# BBM-BBM equations with variable bathymetry\nbbmbbm = BBMBBMEquations1D(bathymetry_type = bathymetry_variable,\n                           gravity = 9.81, eta0 = 0.0)\n\n# Svärd-Kalisch equations with specific parameter set (optimized for small wave numbers)\nsk = SvaerdKalischEquations1D(gravity = 9.81, eta0 = 0.8, alpha = 0.0004040404040404049,\n                              beta = 0.49292929292929294, gamma = 0.15707070707070708)\n\n# Serre-Green-Naghdi equations with variable bathymetry\nsgn = SerreGreenNaghdiEquations1D(bathymetry_type = bathymetry_variable,\n                                  gravity = 9.81, eta0 = 0.8)\n\n# Hyperbolic approximation of Serre-Green-Naghdi equations\nhysgn = HyperbolicSerreGreenNaghdiEquations1D(bathymetry_type = bathymetry_mild_slope,\n                                              lambda = 200.0, gravity = 9.81, eta0 = 0.8)\n                                              # for actual simulations a higher lambda (~500) is recommended\n                                              # it is chosen so low to be able to see the difference between it\n                                              # and the SGN equation.\nnothing # hide\n\nThe initial condition initial_condition_dingemans provided by DispersiveShallowWater.jl automatically sets up the trapezoidal bathymetry and initial wave field.\n\ninitial_co
[…]
