### 2025-11-24 1e7be11
Improve documentation of the analysis callback
