+++ docs/src/callbacks.md
@@ -118,11 +118,13 @@ savefig("analysis_callback.png") # hide
 nothing # hide
 ```
 
+Please note that this plot shows the *change* of the invariants over time, i.e., the difference from their initial values.
 More on different options for plotting the analysis callback can be found in the chapter [Plotting Simulation Results](@ref plotting).
 
 ![analysis callback](analysis_callback.png)
 
-The plot shows that linear invariants such as the total water mass and total velocity are conserved exactly. However, nonlinear invariants such as the entropy may exhibit small growth over time. This occurs because standard time integration methods do not necessarily preserve nonlinear invariants, even when the spatial discretization is conservative.
+The plot shows that linear invariants such as the total water mass and total velocity are conserved exactly. However, nonlinear invariants such as the entropy may exhibit small growth over time (compared to the growth one would obtain from a standard spatial discretization that does not conserve these invariants).
+This occurs because standard time integration methods do not necessarily preserve nonlinear invariants, even when the spatial discretization is conservative.
 
 For a fully discrete entropy-conservative method, see also the following section about relaxation and the [`RelaxationCallback`](@ref).
 
