### 2025-11-24 7dcc176
Update paper/paper.md
Co-authored-by: Hendrik Ranocha <ranocha@users.noreply.github.com>
