### 2025-11-24 a0a37c3
Format plot title in docs
