+++ docs/src/basic_example.md
@@ -120,7 +120,7 @@ plot the solution at the final time by calling `plot` on a `Pair` of the `Semidi
 ```@example overview
 using Plots, Printf
 # default(grid=true, box=:on, dpi=100, titlefont=font(16), linewidth=3, gridlinewidth=2, markersize=4, markerstrokewidth=2, xtickfontsize=14, ytickfontsize=14, xguidefontsize=16, yguidefontsize=16, ztickfontsize=14, zguidefontsize=16, legendfontsize=14) # hide
-plot(semi => sol, plot_title = @sprint "BBM-BBM equations at t = %.2f" last(tspan))
+plot(semi => sol, plot_title = @sprintf "BBM-BBM equations at t = %.2f" last(tspan))
 savefig("shoaling_solution.png") # hide
 nothing # hide
 ```
