### 2025-11-24 7d9d620
state relative change of entropy
