### 2025-11-24 943400a
fix math formatting
