### 2025-11-25 #9361 @Alexander-Barth
@jatkinson1000 The authors have taken all my suggestions into account (https://github.com/NumericalMathematics/DispersiveShallowWater.jl/issues/264). I have now completed my checklist. Let me know if there is more information that is needed from my side.
