### 2025-12-02 ebb354c
Bump the all-github-actions group with 3 updates
Bumps the all-github-actions group with 3 updates: [actions/checkout](https://github.com/actions/checkout), [julia-actions/julia-downgrade-compat](https://github.com/julia-actions/julia-downgrade-compat) and [crate-ci/typos](https://github.com/crate-ci/typos).


Updates `actions/checkout` from 5 to 6
- [Release notes](https://github.com/actions/checkout/releases)
- [Changelog](https://github.com/actions/checkout/blob/main/CHANGELOG.md)
- [Commits](https://github.com/actions/checkout/compare/v5...v6)

Updates `julia-actions/julia-downgrade-compat
[…]
