### 2025-12-02 f24619e
Apply suggestion from @JoshuaLampert
