+++ .github/workflows/Downgrade.yml
@@ -48,7 +48,7 @@ jobs:
           version: ${{ matrix.version }}
       - run: julia -e 'using InteractiveUtils; versioninfo(verbose=true)'
       - uses: julia-actions/cache@v2
-      - uses: julia-actions/julia-downgrade-compat@v2
+      - uses: julia-actions/julia-downgrade-compat@v1
         with:
           skip: LinearAlgebra,Printf,SparseArrays,DelimitedFiles,Test
           projects: ., test
