### 2025-12-02 39b684d
Apply suggestions from code review
