+++ .github/workflows/Downgrade.yml
@@ -60,5 +60,6 @@ jobs:
         with:
           coverage: false
           allow_reresolve: false
+          force_latest_compatible_version: false
         env:
           PYTHON: ""
