+++ versions.js
@@ -1,0 +1,14 @@
+var DOC_VERSIONS = [
+  "stable",
+  "v0.8",
+  "v0.7",
+  "v0.6",
+  "v0.5",
+  "v0.4",
+  "v0.3",
+  "v0.2",
+  "v0.1",
+  "dev",
+];
+var DOCUMENTER_NEWEST = "v0.8.6";
+var DOCUMENTER_STABLE = "stable";
