### 2025-12-10 #9361 @JoshuaLampert
Thanks @jatkinson1000! I double checked the title and author list in the paper and the Zenodo archive as well as the license (MIT license on both the repository and the Zenodo archive).
The current release of the package, which includes the suggestions from the review, is v0.8.6 (released two weeks ago). The DOI for the Zenodo archive is [10.5281/zenodo.10034636](https://doi.org/10.5281/zenodo.10034636).
