### 2025-12-11 #9361 @JoshuaLampert
> [@JoshuaLampert](https://github.com/JoshuaLampert) Please update the affiliations on the zenodo archive to match those on the paper.

I created a PR updating the affiliations of the Zenodo archive in https://github.com/NumericalMathematics/DispersiveShallowWater.jl/pull/274.
