### 2025-12-12 #9361 @kthyng
@JoshuaLampert Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file. For example "drifel'd" but also check others.
