+++ paper/paper.bib
@@ -153,7 +153,7 @@
 }
 
 @article{serre1953contribution,
-  title = {Contribution {\`a} l'{\'e}tude Des {\'E}coulements Permanents et Variables
+  title = {Contribution {\`a} l'{\'e}tude Des {\'e}coulements Permanents et Variables
     Dans Les Canaux},
   author = {Serre, Fran{\c c}ois},
   year = {1953},
@@ -424,11 +424,11 @@
 
 @article{lindeberg2021high,
   title={A High Order Accurate Finite Difference Method for the
-         Drinfel’d--Sokolov--Wilson Equation},
+         {D}rinfel’d--{S}okolov--{W}ilson Equation},
   author={Lindeberg, Ludvig and Dao, Tuan and Mattsson, Ken},
   journal={Journal of Scientific Computing},
   volume={88},
-  number={1},
+  number={18},
   pages={1--22},
   year={2021},
   publisher={Springer},
