+++ paper/paper.bib
@@ -236,7 +236,7 @@
 }
 
 @article{rackauckas2017differentialequations,
-  title={{DifferentialEquations.jl} {--} {A} Performant and Feature-Rich
+  title={{DifferentialEquations.jl} --- {A} Performant and Feature-Rich
          Ecosystem for Solving Differential Equations in {J}ulia},
   author={Rackauckas, Christopher and Nie, Qing},
   journal={Journal of Open Research Software},
@@ -251,7 +251,7 @@
 @article{christ2023plots,
   author={Christ, Simon and Schwabeneder, Daniel and Rackauckas, Christopher
           and Borregaard, Michael Krabbe and Breloff, Thomas},
-  title={Plots.jl --- a user extendable plotting {API} for the {J}ulia
+  title={Plots.jl --- {A} user extendable plotting {API} for the {J}ulia
          programming language},
   journal={Journal of Open Research Software},
   volume={11},
@@ -367,7 +367,7 @@
 }
 
 @misc{klowershallowwaters,
-  title={ShallowWaters.jl - A type-flexible 16-bit shallow water model},
+  title={ShallowWaters.jl --- {A} type-flexible 16-bit shallow water model},
   author={Klöwer, Milan},
   howpublished={\url{https://github.com/milankl/ShallowWaters.jl}},
   doi={10.5281/zenodo.3626102}
