+++ paper/paper.bib
@@ -82,7 +82,6 @@
          smooth solutions},
   author={Giesselmann, Jan and Ranocha, Hendrik},
   year={2025},
-  month={08},
   eprint={2508.04112},
   eprinttype={arxiv},
   eprintclass={math.NA},
